Write docs for subscriptions and reorganize a bit
